refactor: factorize form
