+++ web/src/views/ConsentPortal/OpenIDConnectConsentPortal/OpenIDConnectConsentDecisionFormPreConfiguration.tsx
@@ -1,0 +1,57 @@
+import React, { Fragment, useEffect, useState } from "react";
+
+import { Checkbox, FormControlLabel, Theme, Tooltip } from "@mui/material";
+import Grid from "@mui/material/Grid";
+import { useTranslation } from "react-i18next";
+import { makeStyles } from "tss-react/mui";
+
+export interface Props {
+    pre_configuration: boolean;
+    onChangePreConfiguration: (value: boolean) => void;
+}
+
+const OpenIDConnectConsentDecisionFormView: React.FC<Props> = (props: Props) => {
+    const { t: translate } = useTranslation(["portal", "consent"]);
+
+    const { classes } = useStyles();
+
+    const [preConfigure, setPreConfigure] = useState(false);
+
+    const handlePreConfigureChanged = () => {
+        setPreConfigure((preConfigure) => !preConfigure);
+    };
+
+    useEffect(() => {
+        props.onChangePreConfiguration(preConfigure);
+    }, [preConfigure, props]);
+
+    return (
+        <Fragment>
+            {props.pre_configuration ? (
+                <Grid size={{ xs: 12 }}>
+                    <Tooltip title={translate("This saves this consent as a pre-configured consent for future use")}>
+                        <FormControlLabel
+                            control={
+                                <Checkbox
+                                    id="pre-configure"
+                                    checked={preConfigure}
+                                    onChange={handlePreConfigureChanged}
+                                    value="preConfigure"
+                                    color="primary"
+                                />
+                            }
+                            className={classes.preConfigure}
+                            label={translate("Remember Consent")}
+                        />
+                    </Tooltip>
+                </Grid>
+            ) : null}
+        </Fragment>
+    );
+};
+
+const useStyles = makeStyles()((theme: Theme) => ({
+    preConfigure: {},
+}));
+
+export default OpenIDConnectConsentDecisionFormView;
